+++ digital_twin/environment_interface.py
@@ -4,8 +4,8 @@ Loads an environment from a file
 
 from PIL import Image
 from numpy import ceil
-from environment import Environment, EnvType, get_environment_type_from_value
-from constants import METERS_PER_TILE
+from digital_twin.environment import Environment, EnvType, get_environment_type_from_value
+from digital_twin.constants import METERS_PER_TILE
 
 
 def image_to_environment(width: float, height: float, image_filename: str = "resources/env.png")\
